flowershop/requirements.txt
Django==5.1.7
python-telegram-bot==22.0
python-dotenv==1.0.1
aiogram==3.19.0
dateparser==1.2.1
djangorestframework==3.15.2
Pillow==10.2.0
gunicorn

pytz==2025.1
whitenoise==6.9.0
dj_database_url